# Conflicts:
#	app/src/pitstop/java/com/pitstop/ui/CarScanActivity.java
#	app/src/pitstop/java/com/pitstop/ui/SettingsActivity.java
#	app/src/pitstop/java/com/pitstop/ui/addCarFragments/AddCarUtils.java
#	app/src/pitstop/java/com/pitstop/utils/ServiceRequestUtil.java

+++ app/src/pitstop/java/com/pitstop/models/TripIndicator.java
@@ -22,7 +22,7 @@ public abstract class TripIndicator {
         return rtcTime;
     }
 
-    public int getTripId() {
+    public long getTripId() {
         return tripId;
     }
